+++ ISL.ReIdentification.WebApps.Portals/ISL.ReIdentification.Portals.Client/src/components/csvReIdentification/csvReIdentificationDetailCardView.tsx
@@ -1,22 +1,23 @@
-import React, { FunctionComponent, useState } from "react";
+import React, { FunctionComponent, useState, useRef, useEffect } from "react";
 import { Form, Button, Card, Spinner, Alert, OverlayTrigger, Tooltip, Modal } from "react-bootstrap";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faCircleInfo } from "@fortawesome/free-solid-svg-icons";
+import { reIdentificationService } from "../../services/foundations/reIdentificationService";
+import { AccessRequest } from "../../models/accessRequest/accessRequest";
 import { useMsal } from "@azure/msal-react";
 import UserAccessSearch from "../userAccessSearch/userAccessSearch";
 import { UserAccessView } from "../../models/views/components/userAccess/userAccessView";
 import { OverlayInjectedProps } from "react-bootstrap/esm/Overlay";
-import { useFileUpload } from "../../hooks/useFileUpload";
-import { useCsvData } from "../../hooks/useCsvData";
-import { reIdentificationService } from "../../services/foundations/reIdentificationService";
-import { AccessRequest } from "../../models/accessRequest/accessRequest";
+import { useFileChange } from "../../hooks/useFileChange";
 
 const CsvReIdentificationDetailCardView: FunctionComponent = () => {
+
     const [headerColumns, setHeaderColumns] = useState<string[]>([]);
     const [csvData, setCsvData] = useState<string | null>(null);
     const [selectedHeaderColumn, setSelectedHeaderColumn] = useState<string>("");
     const [selectedHeaderColumnIndex, setSelectedHeaderColumnIndex] = useState<number>(0);
     const [selectedUser, setSelectedUser] = useState<UserAccessView | undefined>();
+    const { submit, loading } = reIdentificationService.useRequestReIdentificationCsv();
     const [error, setError] = useState<string[]>([]);
     const [success, setSuccess] = useState("");
     const [savedSuccessfull, setSavedSuccessfull] = useState(false);
@@ -25,11 +26,15 @@ const CsvReIdentificationDetailCardView: FunctionComponent = () => {
     const [showHelpModal, setShowHelpModal] = useState<boolean>(false);
     const [reason, setReason] = useState<string>("");
     const account = useMsal();
+    const fileInputRef = useRef<HTMLInputElement>(null);
 
-    const { handleFileChange } = useFileUpload(setError, setFileName, setCsvData, setHeaderColumns);
-    const { handleHeaderColumnChange } = useCsvData(csvData, headerColumns, setError, setSuccess);
+    const { handleFileChange } = useFileChange(setError, setFileName, setHeaderColumns, setCsvData, hasHeaderRecord); // Use the custom hook
 
-    const { submit, loading } = reIdentificationService.useRequestReIdentificationCsv();
+    useEffect(() => {
+        if (fileInputRef.current?.files?.[0]) {
+            handleFileChange({ target: fileInputRef.current } as React.ChangeEvent<HTMLInputElement>);
+        }
+    }, [hasHeaderRecord]);
 
     const handleSubmit = (e: React.FormEvent<HTMLFormElement>) => {
         e.preventDefault();
@@ -81,8 +86,35 @@ const CsvReIdentificationDetailCardView: FunctionComponent = () => {
         });
     };
 
+    const handleHeaderColumnChange = (e: React.ChangeEvent<HTMLSelectElement>) => {
+        const selectedColumn = e.target.value;
+        setSelectedHeaderColumn(selectedColumn);
+        const index = headerColumns.indexOf(selectedColumn);
+        setSelectedHeaderColumnIndex(index);
+
+        if (csvData) {
+            const decodedData = atob(csvData);
+            const rows = decodedData.split("\n");
+
+            if (rows.length > 1) {
+                const nextRow = rows[1].split(",");
+                const value = nextRow[index];
+
+                if (/^\d{10}$/.test(value)) {
+                    setError([]);
+                    setSuccess(`The value "${value}" in the next row at the selected column index is 10 digits long and is a valid Pseudo Identifier.`);
+                } else {
+                    setSuccess("");
+                    setError([`The value "${value}" in the next row for the selected column index is not a valid Pseudo Identifier, please follow the guidance in the help section.`]);
+                }
+            }
+        }
+    };
+
     const handleCheckboxChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-        setHasHeaderRecord(e.target.checked);
+        const newValue = e.target.checked;
+        setHasHeaderRecord(newValue);
+        setError([]);
     };
 
     const renderTooltip = (props: OverlayInjectedProps): React.ReactElement => (
@@ -104,6 +136,7 @@ const CsvReIdentificationDetailCardView: FunctionComponent = () => {
                     </Card.Header>
 
                     <Card.Body>
+
                         <Card.Subtitle className="text-start text-muted mb-3">
                             <small>
                                 Please upload your CSV file below, specify the reason for identifying these patients
@@ -131,7 +164,9 @@ const CsvReIdentificationDetailCardView: FunctionComponent = () => {
                                         onChange={handleFileChange}
                                         placeholder="Upload CSV"
                                         accept=".csv"
-                                        required />
+                                        required
+                                        ref={fileInputRef} />
+
                                 </div>
                                 <Form.Text className="text-muted">
                                     Please upload your CSV (other file types will be rejected).
@@ -144,11 +179,7 @@ const CsvReIdentificationDetailCardView: FunctionComponent = () => {
                                         <Form.Label><strong>Select Identifier Column From Csv:</strong></Form.Label>
                                         <Form.Select
                                             value={selectedHeaderColumn}
-                                            onChange={(e) => {
-                                                setSelectedHeaderColumn(e.target.value);
-                                                setSelectedHeaderColumnIndex(headerColumns.indexOf(e.target.value));
-                                                handleHeaderColumnChange(e);
-                                            }}
+                                            onChange={handleHeaderColumnChange}
                                             required>
                                             <option value="" disabled>Please select a column...</option>
                                             {headerColumns.map((column, index) => (
@@ -166,6 +197,7 @@ const CsvReIdentificationDetailCardView: FunctionComponent = () => {
                             )}
                             <Form.Group className="text-start">
                                 <Form.Label><strong>Re-identification Reason:</strong></Form.Label>
+
                                 <Form.Control
                                     as="textarea"
                                     value={reason}
@@ -174,6 +206,7 @@ const CsvReIdentificationDetailCardView: FunctionComponent = () => {
                                     rows={3}
                                     required
                                 />
+
                                 <Form.Text className="text-muted">
                                     Please supply a reason why you are requesting to re-identify, this will be visible in the email to the recipient.
                                 </Form.Text>
@@ -209,6 +242,7 @@ const CsvReIdentificationDetailCardView: FunctionComponent = () => {
                         <p>Alternatively, the recipient can launch their worklist in the portal and re-identify from there.</p>
                         <p><strong>Note:</strong> Files will last for 7 days and then be removed from the system.</p>
                     </Alert>
+
                 </>
             )}
 
@@ -243,6 +277,6 @@ const CsvReIdentificationDetailCardView: FunctionComponent = () => {
             </Modal>
         </>
     );
-};
+}
 
 export default CsvReIdentificationDetailCardView;
